requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
pydantic
python-multipart
jinja2